+++ app/NivTest.hs
@@ -1,0 +1,6 @@
+module NivTest where
+
+import Niv.Test
+
+main :: IO ()
+main = Niv.Test.test
