--- app/test.hs
@@ -1,8 +1,0 @@
-module Main where
-
-import Niv.Test
-import qualified Test.Tasty as Tasty
-
-main :: IO ()
-main = Tasty.defaultMain $ Niv.Test.tests
-
