+++ default.nix
@@ -35,7 +35,7 @@ with rec
         ''
           repl() {
             shopt -s globstar
-            ghci -Wall app/**/*.hs src/**/*.hs
+            ghci -Wall app/NivTest.hs src/**/*.hs
           }
 
           echo "To start a REPL session, run:"
